+++ typescript/discovery_kit_api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## UNRELEASED
+
+ - Support target enrichment rules.
+
 ## 1.1.0
 
  - Support restriction of discoveries to AWS agents.